No help@kbase
(There isn't a way to make line-by-line comments on a file in GitHub, only on a PR, so I had to edit this file rather than commenting on it, which is what I wanted to do.)

"Issues in non-production environments can be reported to help@kbase.us" --> NO.
But I'm not sure what the right thing is here--create a ticket in JIRA, ask on Slack (what channel?)? And the Help Desk isn't responsible for assigning internally-created tickets, just external ones.

Also, what is "pagerduty"?